% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSystem.R
\name{reportSystem}
\alias{reportSystem}
\title{report the system}
\usage{
reportSystem(hypothesis = braw.def$hypothesis, design = braw.def$design)
}
\value{
ggplot2 object - and printed
}
\description{
report the system
}
\examples{
reportSystem(hypothesis,design)
}
